feat: Moved logic from new_molecule to molecule
