+++ src/manim_chemistry/twoD/molecule.py
@@ -1,3 +1,5 @@
+from typing import Optional, Tuple, Dict
+
 from manim import (
     DOWN,
     GREEN,
@@ -11,6 +13,8 @@ from manim import (
 )
 
 from ..utils import mol_parser, mol_parser_string, sdf_parser, sdf_parser_string
+from ..manim_chemistry_molecule import MCMolecule
+
 from .atom import MAtomObject
 from .bond import *
 
@@ -20,7 +24,7 @@ class MMoleculeObject(VGroup):
         self,
         atoms_dict: dict,
         bonds_dict: dict,
-        representation_type: str or None = None,
+        representation_type: Optional[str] = None,
         explicit_carbons: bool = False,
         explicit_hydrogens: bool = False,
         planar: bool = True,
@@ -63,7 +67,6 @@ class MMoleculeObject(VGroup):
             )
             atoms.add([(index, matom)])
             atoms_by_index[index] = matom
-
         return atoms, atoms_by_index
 
     def get_bonds(self):
@@ -282,6 +285,55 @@ class MMoleculeObject(VGroup):
                                 atom.element + "H", bonds_direction
                             )
 
+    @staticmethod
+    def molecule_from_file(filepath, *args, **kwargs):
+        """
+        Reads a file and returns a single molecule from that file.
+
+        By default retrieves the first molecule generated by the file.
+
+        Args:
+            filepath (_type_): Pathlike or str
+            label (bool, optional): Add a label such as element symbol or number. Defaults to False.
+
+        Returns:
+            MMoleculeObject: GraphMolecule from the file
+        """
+        mc_molecule = MCMolecule.construct_from_file(filepath=filepath)
+        if isinstance(mc_molecule, list):
+            mc_molecule = mc_molecule[0]
+
+        atoms, bonds = MMoleculeObject.mc_molecule_to_atoms_and_bonds(mc_molecule=mc_molecule)
+
+        return MMoleculeObject(atoms, bonds, *args, **kwargs)
+
+    @staticmethod
+    def multiple_molecules_from_file(filepath, *args, **kwargs) -> VGroup:
+        """
+        Reads a file and returns a collection of molecules from that file as a VGroup.
+
+        Args:
+            filepath (str | Pathlike): Path to the molecule
+
+        Raises:
+            Exception: In case the mc_molecules parsed is not a list.
+
+        Returns:
+            VGroup: VGroup with the molecules inside.
+        """
+        mc_molecules = MCMolecule.construct_multiples_from_file(filepath=filepath)
+
+        if not isinstance(mc_molecules, list):
+            raise Exception(f"Expected a list of molecules. Received {mc_molecules}")
+
+        mmolecules = VGroup()
+        for mc_molecule in mc_molecules:
+            atoms, bonds = MMoleculeObject.mc_molecule_to_atoms_and_bonds(mc_molecule=mc_molecule)
+            mmolecules.add(MMoleculeObject(atoms, bonds, *args, **kwargs))
+
+        return mmolecules
+
+
     def from_mol_file(filename, *args, **kwargs):
         atoms, bonds = mol_parser(filename)
         return MMoleculeObject(atoms, bonds, *args, **kwargs)
@@ -466,6 +518,57 @@ class MMoleculeObject(VGroup):
 
         return bonds_positions
 
+    @staticmethod
+    def mc_molecule_to_atoms_and_bonds(mc_molecule: MCMolecule) -> Tuple[Dict, Dict]:
+        """
+        Transforms the structure of a mc_molecule to a (vertices, edges) tuple
+        with the following structure:
+        - Vertices: {<atom_index>: MCAtom}
+        - Edges: {(<from_atom_index>, <to_atom_index>): MCBond}
+
+        Args:
+            mc_molecule (MCMolecule): _description_
+
+        Returns:
+            Tuple[Dict, Dict]: _description_
+        """
+        atoms = {}
+        for index, atom in mc_molecule.atoms_by_index.items():
+            bond_to = {}
+            for bond in atom.bonds:
+                #TODO: This patches an issue with from_atom and to_atom but does not solve it completely. 
+                # Fix later the root cause
+                if bond.from_atom.molecule_index == index:
+                    bond_to[bond.to_atom.molecule_index] = bond.to_atom.element.symbol
+
+                else:
+                    bond_to[bond.from_atom.molecule_index] = bond.from_atom.element.symbol
+
+            atom_data = {
+                "coords": atom.coords,
+                "element": atom.element.symbol,
+                "bond_to": bond_to
+            }
+
+            atoms[index] = atom_data
+
+        bonds = {}
+        for atom_index in atoms.keys():
+            atom_bonds = []
+            for bond in mc_molecule.bonds:
+                if atom_index == bond.to_atom.molecule_index:
+                    atom_bonds.append({
+                        "to": bond.from_atom.molecule_index,
+                        "type": bond.bond_type,
+                        "stereo": bond.stereo,
+                        "topology": bond.topology,
+                        "reacting_center_status": bond.reacting_center_status
+                    })
+            if atom_bonds:
+                bonds[atom_index] = atom_bonds
+
+
+        return atoms, bonds
 
 class NamedMolecule(VGroup):
     def __init__(
